added seqs to regex converter
