+++ pyrepseq/__init__.py
@@ -2,6 +2,7 @@ from .io import *
 from .distance import *
 from .stats import *
 from .nn import *
+from .util import *
 from .clustering import *
 from . import plotting 
 from .version import __version__
